requirements.txt
# Core API 
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# PostgreSQL 
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Authentication 
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# RAG / Embeddings
torch==2.4.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
packaging>=23.0
sentence-transformers==3.1.1
transformers==4.45.2
numpy==1.26.4

# Mistral AI 
mistralai==1.2.2

# Vector Database 
qdrant-client==1.11.3

# Retrieval 
rank-bm25==0.2.2

# Utilities 
python-dotenv==1.0.1
tqdm==4.66.5
httpx==0.27.2
orjson==3.10.7
typing-extensions==4.12.2
firebase-admin==6.5.0

#Document_summarizer
PyMuPDF==1.24.5
python-docx==1.1.2
python-multipart==0.0.9

